Rewritten WRAM, VRAM, PRGROM, CHROM and Nametebles management. (#308)
- Added support for nes20db.xml and dip.cfg (many thx to NewRisingSun for is for his immense work).
  Both files will be distributed together with the emulator and will be located in the following directories:
  * Linux/BSD - {default system datadir}/puNES
  * Windows - directory where the executable is placed.
  if you want to use a more updated version than the one distributed with the emulator, just copy it either in the executable directory or in the working directory of the emulator, which can also be opened from the "File -> Open working directory" menu.
- Added fully support to NES 2.0 file cart format header.
- Fixed a small bug in the evaluation of sprites. Thanks to Quietust's read2004.nes test rom.
- Rewritten mappers : All.
- Rewritten the format and management of save states.
  WARNING save states of version 0.110 or earlier are no longer compatible.

Note: Due to the numerous changes I'm making to the core of the emulator, new bugs may have been introduced, if you find roms no longer work correctly compared to the previous version, please let me know.

+++ src/core/bck_states.h
@@ -34,10 +34,10 @@ enum bck_states_operations_mode {
 #define EXTERNC
 #endif
 
-EXTERNC void bck_states_op_screen(BYTE mode, void *data, size_t *index, size_t *size_buff);
-EXTERNC void bck_states_op_keyframe(BYTE mode, void *data, size_t *index, size_t *size_buff);
-EXTERNC void bck_states_op_input(BYTE mode, void *data, size_t *index, size_t *size_buff);
-EXTERNC void bck_states_op_input_port(BYTE port, BYTE mode, void *data, size_t *index, size_t *size_buff);
+EXTERNC void bck_states_op_screen(BYTE mode, void *data, size_t *index, uint64_t *size_buff);
+EXTERNC void bck_states_op_keyframe(BYTE mode, void *data, size_t *index, uint64_t *size_buff);
+EXTERNC void bck_states_op_input(BYTE mode, void *data, size_t *index, uint64_t *size_buff);
+EXTERNC void bck_states_op_input_port(BYTE port, BYTE mode, void *data, size_t *index, uint64_t *size_buff);
 
 #undef EXTERNC
